requirements.txt
pyautogui
opencv-python
pillow
pynput
mss
